+++ base/urls/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from base.views import views
+
+urlpatterns = [
+    path('', views.getRoutes),
+]
